Refactor opendss.Storage and implement full API
Add simple implementation of StorageDevice.get_kw_rated and
StorageDevice.kwh_rated.

Extend the dssutil API with a function to query properties (this
is needed since run_command will raise an error if the command
has a return value. In the new function we look at the return
value to make sure is doesn't have an error message inside it,
and return it otherwise).

Refactor opendss.Storage.set_power() to simplify it. The
Storage._edit() method is replaced with Storage._set()
which is used to set a single property at a time.