Publish storage power as a global publication
In order to support multiple storage devices controlled by the
same controller federate without needing to pass the names of
the various controller federates to the grid federate we make
the power set-point for the storage device a global publication.
Since it is global the grid federate can subscribe to it just
based on the storage device name.