Set stepsize to the actual time-delta before solving
OpenDSS updates the state of charge based on the stepsize
times the number of solutions. When we use irregular step
sizes we must explicitly set the step size (before setting
the clock) to the time that has elapsed since the last
solution. Because the step size is no loner consistent we
must also add some code to explicitly set the time of the
next request to the maximum step size (currently set at 15
minutes, but should be made a parameter).