Add grid.GridSpecification class
First cut of a class that provides a specification of the grid
and the connected devices. The specification can be used to
construct an OpenDSS model (an instance of ssim.opendss.DSSModel).

It is designed to be independent of any particular model, providing
just a basic description of the grid that can be used to initialize
the simulation federates. It is also intended to be serializable so
it can be saved to a file or, most importantly, shared between
processes.

In addition to grid.GridSpecification, a grid.StorageSpecification
dataclass is added which specifies the properties of storage
devices that are to be added to the grid model. I intend for this
to be the prototypical way to add extra devices to the grid
specification. A dataclass with some common parameters and a
dictionary of additional parameters that are used by the grid model
or federates. This can be seen in action in the
DSSModel.from_grid_spec() method which constructs a dictionary of
OpenDSS parameters for storage objects by combining the common
parameters (kw_rated and kwh_rated) with the extra parameters
dictionary.

The GridSpecification API will likely need to grow and adjust
as it is used to share data with a wider variety of federates;
however, it should be maintained as nothing more than a
specification of the grid and the devices connected to it.
It is tempting for me to try to shoe-horn in some capability to
specify which federates should be responsible for which externally
controlled devices, or which additional values should be published
by the grid federate. These should, I think, be handled separately
in order to keep the grid specification as simple and broadly
useful as possible. An additional "federate specification" may be
added later to provide this capability if/when it is needed.

* Add simulator.Simulation() class

The Simulation class can be used to initialize and run a
simulation from a GridSpecification. This is a first cut
that needs to be updated to support additional/different
numbers of federates. The existing simulator.run_simulation()
function has been rewritten to demonstrate the use of the new
class.

Some of the federate initialization has been consolidated into
functions in the ssim.simulator module to make is more straight-
forward to add new federates and reduce the amount of repeated
boiler-plate HELICS code.