Add controller, controller_params fields to StorageSpecification
The controller is now initialized by passing the full device spec
to the constructor. I separated the device params and the
controller_params dictionaries in StorageSpecification just for
organization, it might be worth combining them since some
controllers may use both, but I like the semantics of two separate
fields.

IdealStorageModel still needs to be updated to take the device
spec as the only argument to its constructor.