First cut of grid.PVSpecification dataclass
PV Systems that have been added to the grid specification
passed to DSSModel.from_grid_spec() are added to the OpenDSS
model. As a first cut many OpenDSS PVSystem object params
are assumed to be present in the PVSpecification.params dict.

There are a number of curves that need to be specified for
the PVSystem object, Two fields have been added to the
PVSpecification class as a starting point for implementing
these; however, these fields are not yet used. The full
implementation of the grid-spec -> dss model api is going to
require a specification of OpenDSS "Loadshape" and "XYCurve"
objects.

Additionally, I think the DSSModel.add_pvsystem() API should
be re-thought to more closely resemble DSSModel.add_storage(),
with a few key parameters and a dictionary of extra parameters.