Add efficiency and PT curves for PVSystems
If the efficiency or pt curve fields of PVSpecification are
not None, then a new XY-curve is added for the data in each
field. The curve is assigned to the respective parameter of
the OpenDSS PVSystem object. Each curve in the opendss model
is named according the type of curve ("eff" or "pt") and the
name of the PV system.

If many PV systems are added and each one has a pt_curve or
an inverter_efficiency curve specified, many separate XYCurves
will be created in OpenDSS. If the curves are all the same, or
if many of them are the same, this is not particularly
efficient. A better solution may be to consolidate the curves
for each PVSystem (and possibly storage devices) and point
the systems that share common curves to the same XYCurve
object. Not really worth it at the moment, but may be
important if memory usage becomes a concern.