Update DSSModel.solve() to make time a required param
Again, because the stepsize is no longer a fixed reference
point inside opendss, we can't solve the opendss model
without manually setting the opendss time. Doing so would
result in the step size slowly creeping smaller and smaller
until the simulation grinds to a halt.

Refactor slightly by moving the stepsize adjustments into the
method for setting the OpenDSS clock. Makes more sense to me
that these operations should be in the same place since every
time we set the clock we must be sure to adjust the step size
first.