First cut of a reliability model added in ssim.reliability
Adds a simple API for reliability models to get the time of the
soonest event and to iterate over all events that occur at (or
before) a given time. My vision is that this API can provide access
to a SimPy simulation of the grid reliability. I don't think it is
quite there in its functionality, but it will serve to get the
communication topology between the reliability federate and the
grid federate up and running.

Adds a data structure to be used for representing events (events
can be either failures or restorations) that can be serialized as
a JSON string. This data structure will support generic
'reliability' messages between the reliability federate and the
grid federate.

At this point the model should be functional and can be plumbed in
to the simulation via a new federate. As part of that, though we
should consider how to initialize the reliability model. For now,
I can pass the grid specification to the federate, and it can load
the opendss model to get the list of lines (the only component with
a ReliabilityModel implementation, at the moment). From the list of
lines a collection of LineReliability objects can be created and a
GridReliabilityModel can be instantiated with the line models as
its components. The federate will use `GridReliabilityModel.peek()`
to determine its next time request.