First cut of reliability federate
Initial implementation of a federate for the ssim.reliability
reliability model. Includes a single endpoint in the grid
federate where all reliability messages are sent.

Still needs to be hooked in to the Simulator class to run as
part of the simulation. Also need to pass model parameters
for each component/the whole grid (for now just have fixed,
unrealistic values to force frequent failures while I'm testing
it).

Also need to add support for failures in other types of components,
but that is beyond the scope at the moment. Only simulating failure
of Lines.