Run the reliability federate
Spin up a process for the reliability federate and run it along
with the rest of the simulation.

A few bug fixes in both the reliability federate (sending messages
correctly), the grid federate (applying resilience events
correctly), and the GridReliabilityModel (using itertools.chain
correctly).

Still need an API for configuring the reliability federate/model,
currently just passing the names of two lines from the IEEE13 bus
circuit in _init_reliability_federate. I will need to put some
concerted thought in to how to configure this in a reasonable and
modifiable way (I doubt we will get it right on the first try).