Add API to ssim.opendss.DSSModel for line/switch failures
In OpenDSS switches are just special lines, so this API works
for both elements. A line can be failed open, closed, or current.

When a line is failed any switch controllers associated with the
are locked to prevent normal operation. If/when an API is added
for "normal" switch operations extra care will need to be taken
that the switch controls have been restored to a functional state.
For now, since no such API exists, it suffices to ensure that the
switch controls are unlocked after the line is restored.