Re-integrate logger federate
Adds logger federate back to the federation and configures
it with a JSON file. The JSON is currently only used to set
up the federate options, and not the subscriptions. This
was the shortest path to something that works, but I think it
is also a reasonable approach.

Some re-naming in the Voltage logger is required, since we
now subscribe to topics named for the storage devices, not the
busses of interest.