Apply storage controller parameters from grid config file
Adds a method to look up storage devices by name in the
GridSpecification. Not implemented efficiently, but also not
a bottleneck (should only be called once, and the list of
devices should not be too large). If too slow can be easily
adapted in the future to use a dict or bisect.