Use opendssdirect.run_command() to set OpenDSS time
Generally the correct approach for setting time (or interacting
with OpenDSSDirect in any way) is to use the API functions rather
than relying on run_command(). By using the API functions we get
error checking and (potentially) better performance. In this
case, however, the API functions are leading to incorrect results
when running a simulation in "time" mode with irregular timestep
spacing. Under these conditions, anomalous "spikes" appear in
the voltage and total power at apparently random times. By using
a text-based command to set all three values (hour, second, and
stepsize) simultaneously we can eliminate the spikes and get the
expected results from the simulation.

In addition to the change in how time is set, this commit
simplifies the calculation of the hour and seconds values by
using div (//) and mod (%). Originally, I must have been
thinking that python required integers for these operations, but
it turns out they both operate on floating point values. (//)
truncates the fractional part of its output and (%) adds the
fractional part to its output which is precisely the behavior we
need.