Initial implementation of EMS federate and EMS model
Adds and EMS model which dispatches storage devices based on
a fixed time window (i.e. discharge in the on-peak time and charge in the off-peak time).

Still needs to be connected with a forecast (for more
sophisticated dispatch) and to the reliability model to make
dispatch decisions that are informed by the current state of
the grid.