Send reliability events to both the grid & ems federates
Since there are only two endpoints that need to receive the
reliability event message I just send it twice. In principle
we should be able to use a HELICS cloning filter to do the
same thing, but this resulted in some unexpected HELICS errors
(federate disconnected). Others on the HELICS discussion board
mentioned that they had not been successful using cloning
filters either, so I took the somewhat ad-hoc approach of
sending two separate messages.

I'm not sure of the performance implications of this, but
with only two receiving endpoints, I suspect it is about as
performant as the cloning filter. If more endpoints are need
to receive the reliability messages, we will want to re-visit
this.