Add grid.PVStatus dataclass
This class holds the data sent from the grid federate
to the EMS federate when reporting the current status
of the PV systems connected to the grid.

By defining it as a data class in ssim.grid we can keep
the representation of the data sent between the federates
independent of the underlying grid and EMS model. We
add a similar class for the messages sent by the storage
controller federate, which are currently implemented as
raw dicts. Hiding the serialization and message format
might save us some headaches in the future.