Refactor ssim.reliability
Previous reliability model was cumbersome and difficult
to extend or modify. The objective with this refactoring
is to provide a single generic class for a component's
failure model (MultiModeReliabilityModel). This model is
populated with multiple failure mode models that capture
the behaviour of the original reliability model, but can
be instantiated with component-specific parameters rather
than requiring separate implementations for each component.

There will be separate implementations required for each
general class of failure mode, for now only an AgingFailure
is implemented. To support generator failures we will add
an OperatingTime failure mode as well. This provides the same
flexibility, allowing specialized implementations for
individual components, but improves on the previous
implementation by supporting better re-use of generic models
such as aging related failure.

New classes are added to represent failure and restoration
events internally. The new Failure event includes the repair
event that will be triggered when the repair time has elapsed.
This will facilitate failure modes that include complex models
with repair events that depend on the exact nature of the
failure.