Add wear-out model for generators
This is the first reliability model that has two failure
modes. The first is the same aging-type failure that was
implemented for lines and swtiches. The second is an
operating wear out, based on the opertating time for the
generator. This forced some tweaks to the multi-mode
reliability model. Specifically, the individual failure
modes are not updated after they have added failures to
the pending failure queue. Once they are removed they are
reset, whereupon they sample a new failure and are again
updated as time advances.