Enable and disable elements connected to grid nodes
Changes the names of the keys used for each element
to match the opendss names exactly. This makes it easier
to look up elements based on the full opendss name, and
makes the code easier to read by elminating the need for
long chains of if/elif conditional statements to match
element types. Instead the correct device set can be
looked up directly by the element type string.