Add loadshapes for loads and updated irradiance profile
Adds a 5-minute load shape and a 5-minute irradiance profile (clearsky
GHI).

When adding the irradiance profile loadshape to the grid model, the file
is first read to count the number of points. This gives us more
flexibility in configuring the irradiance profiles; however, it relies
on the assumption that the load-shape csv file includes timestamps as
well as irradiance values.