Record loading percent at all power delivery elements
This defaults to all elements, which is fine for the 13
bus feeder, but to keep it efficient with larger systems
we need to be able to specify a sub-set of PDE to monitor.

The ability to monitor a subset is implemented, but the
it is not exposed through the config file yet. We will need
to add a list of PDE to monitor in the config file and pass
that to the grid model through the gridspec structure.