Refine ems.GridModel and rough out a composite EMS
This is not complete, needs more work to operate the
EMS from the ems federate and more work to dispatch
storage inside the EMS. Also need work to implement
the dispatch of generators connected to the grid.

Does not implement the communication layer. Need to merge
that in from the grid-ems-comms branch once this is
complete.