Send status of every load to the EMS federate
Add an endpoint for loads in the grid federate ("load.control")
from which status messages are sent.

Add LoadInterface class to the grid federate to manage the
load.control endpoint. This just publishes status messages, but could
be extended in the future to handle external changes to
loads. Currently only one endpoint serves all loads, but it may be
reasonable to split this into one endpoint per load object. Before
doing that, we should look in to the scalability of both
approaches (it is not clear whether one endpoint with 10000 messages
scales better than 10000 endpoints with 1 message each).

Add grid.LoadStatus dataclass.

Add loads to EMS grid model (so load values can be associated with
specific grid nodes & components).