Return control messages from all component EMSs
For each component EMS in the CompositeHeuristicEMS yield every
control message. Required some changes to match the case convention of
storage device names (lower case) and some updates to the data
structures in the HeuristicEMS.

Revealed a bug that needs to be fixed where discharge dispatch
messages are always followed immediately by an "idle" message.