Make storage control endpoint lower case
The name of the storage device in the control endpoint is made
lower case to match the assumptions in the ems federate.

The EMS federate assumes that the names of grid components are not
case sensitive. We might want to revisit this since it is based on the
naming in OpenDSS and may not hold if we changed to a different grid
simulator.