Move status message parsing and serialization into the grid module
To simplify parsing and make it easier to add new messages in the
future I have added a StatusMessage base class from which all of the
status message dataclasses inherit. The base class provides a single
method for serializing all status messages to JSON strings and adds
information about the message type to the JSON representation which
will simplify parsing and removes the requirement that you must know
the source of the message in order to parse it. In addition to the
serialization method, the base class has a classmethod which works as
a factory when de-serializing JSON into concrete message instances.

Specify __slots__ for all status message data classes

By specifying slots, we can substantially reduce the memory overhead
for each status message. This can become important as we scale up to
larger grid models, particularly for load messages since the number of
loads will most likely scale linearly with the number of busses in the
grid. This change required removing the default value for
GeneratorStatus.online.