Use frozenset() to represent components in EMS dictionaries
Previously I converted the sets of nodes that form connected
components into tuples in order to make them hashable; however, this
was an error and introduced a bug beacuse the order of the elements in
the tuple is unspecified (since sets are not ordered). To avoid the
potential bugs caused by the same set represented as a tuple of a
different order we can instead use the frozenset() type. Frozen sets
maintain the semantics of sets (equality determined by membership)
while resulting in an immutable (and therefore hashable) object that
can be used as a dictionary key.

This commit also removes the inverted indexes created in the
CompositeHeuristicEMS.update() method, replacing them with calls to
GridModel.component_form_element(). I'm not sure this is a performance
improvement, but it substantially improves the readability of the code
and provides a layer of separation that can provide an opportunity for
performance improvements if they are needed in the future.