Act on GeneratorControlMessages received in the grid federate
Add ssim.opendss.Generator.change_setpoint() method to support setting
the generator output in a way that respects whether or not the
generator is in a failed or operational state. This is a pretty
rudimentary API at this point, but it can be expanded to support more
complex failure modes (such as curtailed maximum power output) in a
way that will keep all the changes isolated inside the ssim.opendss
module.

With the message object, and the response to the message object,
implemented the EMS still needs to send generator control messages to
the grid federate. To do that it will need to know which generators
can be dispatched, and the specs of those generators (primarily
minimum and maximum real/reactive power output). These will need to be
added in the grid config file.