Add optional ems key to grid configuration
Adds a dataclass to represent the ems configuration and an attribute
`GridSpecification.ems` which is assigned to the EMS configuration, if
provided.

The EMSSpecification class is very simple initially, containing an
attribute to be used for indicating what kind of EMS to use, and an
attribute to be used for a dict containing parameters needed to
configure the EMS instance. I think this will probably be sufficient
over the long term, but if we want to have different EMS instances
managing different microgrids, then we will need to do something
different.