Add generator for a sequence of times granted by HELICS
Abstracts the common pattern in all federates and provides uniform
logging of the times that are granted, whether the time grant is a
preemption, and which inputs/endpoints triggered the preemption.

Still needs some work with respect to logging the updated
inputs/endpoints. Periodically the storage federate is preempted
apparently without any enpoint being updated. This doesn't really make
sense, based on how I understand preemption to work, so I suspect an
error in the logging.