Make next_update param of timing.schedule optional
The impact of this is to simplify the use of timing.schedule() for
federates that are purely event-driven and do not make any time
requests other than to sleep until pre-empted or until all other
federates have finished.

Also correct a bug (typo) in federates.ems and avoid HELICS
deprecation warnings in federates.timing by using the dict keys for
input and endpoint names.