Ensure subscriptions and enpoints are checked in log_preemption
Previously for combination federates only the subscriptions were
checked for updates when logging a preemtion.

Made some changes to the log statements as well in the interest of
understandability, including removing the the time stamp that was
granted when the federate is preempted since HELICS3 includes that
information in the log metadata output.