Add ssim.opendss.Storage.state_change() method
This method returns the time until the storage device is depleted or
fully charged, accounting for the loss and efficiency model used by
OpenDSS.

This method will be used to prevent the grid federate from advancing
its clock past significant events related to changes in storage power
input/output. It could also be used to provide additional information
in the StorageStatus messages sent to the EMS.

The ssim.opendss.Storage.status property is updated to correct a bug
in construction of the status messages (was setting fields that were
removed from the message class).