Set the storage update time to inf if there are no ESS on grid
Without this guard, the min(...) call will raise a ValueError when the
set of storage devices is empty. Using inf in the case where there are
no ESS ensures that the other min(...) calls will always return the
next time from the control queue or other actions when there are no
ESS connected to the grid.