Configure storage-related pubs & subs using API not JSON
To streamline the configuration of additional storage devices, this
commit changes the storage federate config file by removing all
options that depend on the name of the storage device (federate name,
subscriptions from the grid federate) and instead configuring those at
runtime based on the name of a storage device passed on the command
line. The grid federate is also updated to configure its pubs and subs
based on the storage devices in the grid specification, rather than
requiring the publications for soc, power, and voltage be configured
in the federate config file.