Include federate config files in the ssim package
Adds [options.package_data] section to setup.cfg and changes from a
static list of packages to 'find_namespace:'

Moves federate configs inside the ssim.federates package. This is the
prefered way to package data files for distribution. It also has the
benefit of keeping the code and the configurations in the same place.