First cut generate and evaluate grid configurations
The evaluation step will need to be considered more carefully when we
are integrating with the quest gui so we can poll or be notified when
the simualtor has finished. subprocess.Popen objects support this, so
we should just need to find the right way to use it.

Includes a function to turn a grid.StorageSpecification into a dict so
we can use the same object in the generator as in the simulator. That
way we don't have duplication of how specific storage configurations
are represented.