Refactor storage configuration screen
Substantially refactors the storage configuration screen to take
advantage of kivy properties and to abstract the concept of a list
view with items that can be deleted.

The attributes of the storage device that is being configured are now
stored in class members of the screen and the StorageOptions instance
is not created until the save button is pressed. This enables better
separation between the UI logic (redrawing lists, etc.) and the objects
used to configure the simulation.

This is still not complete. The main additional requirement is to
prevent adding storage elements with duplicate names.