Don't set the storage state after setting a new power setpoint
Previously we needed to explicitly set the state of the storage device
to idle, charging, or discharging after setting a new power
setpoing. In the newest version of OpenDSSDirect the state is changed
automatically based on the value of the power setpoint. Setting the
state manually based on the Storage.kw value will cause the power
setpoint to be changed to the maximum charge or discharge rate,
depending on the state that is set.