Use POSIX-style paths in config files even on Windows
HELICS CLI can't handle Windows style paths, so we work around this
by converting all our paths to `pathlib.PurePosixPath` before writing
the JSON files.

There are also a few changes to the command line arguments for the
federates to fix some missing args, in particular to the metrics
federate.