Use importlib_resources to load Kivy-related resource files
This means that the images, fonts, and kv files can be packaged by the
application in any python ditribution format.

There are three lists for file names that need to be maintained:
fonts, images, and kv files. These need to be kept as lists of files
so that importlib_resources can extract and write the files into a
temporary location when the application is packaged as a zip
file (e.g. a wheel or other bdist).