Add MetricManager.to_dicts() method
Convert a MetricManager instance to a list of dicts that correspond
with the JSON structures required for configuring metrics in the grid
configuration file.

The MetricAccumulator.to_dict() method is added to support this.
Because accumulator names are not stored as an attribut of the
MetricAccumulator class, this method must take the name as an argument
passed by MetricManager.to_dicts()