Simplify classes for reliability param forms
Reduce the number of classes and layers by taking advantage of the
fact that tabs inherit from a layout.

Implement a common interface for validation and saving across
relibility form classes which now all inherit from
`ReliabilityModelTab`

Still needs work on validation which is currently rudimentary at
best.

Validation of relationships between parameters needs to be handled. It
is currently possible to set the max repair time to a value less than
the min repair time - might be able to fix this by dispatching an
event to the max field whenever the min changes (and vice-versa) to
adjust the min/max values in the input validation.