Use grid model fingerprint instead of path in Project.__hash__()
This means that if two Project instances use the same grid model, but stored at
different paths, they will still share the same hash. The __hash__ method will
fall back on the grid model path if a model instance is not available (e.g. if
the path does not exist).