Fix grid model path errors when running checkpoints
By using an absolute path when we write the grid JSON file we can avoid problems
where the relative paths when loading the grid model don't work out. This comes
at the expense of making the configurations portable, but I think it is a
reasonable trade off because the JSON doesn't need to be portable, just the grid
model and project.toml file written by the checkpoint.