StorageControl class wraps InverterControl
This commit provides the complete implementation of the refactoring begun in the
prebious commit. Details of how the InverterControl class is used are fleshed
out and the rest of the interface defined by InverterControl is implemented on
StorageControl. Some tests have been added mainly to facilitate debugging. They
are not even close to adequate, but they have been very helpful.