Import matplotlib before anything else happens
This is a suggestion from Kevin Anderson who was able to reproduce
the crash and found that it was related to the order of imports in
ssim.federates.logger. He found that importing matplotlib first, even in a separate session appears to fix the issue. Unfortunately his reproducer doesn't work on my machine, but I'm going to give his 
suggestion a shot and hope it fixes it.

If it doesn't I'll go modify federates.logger to import mpl first.