Use HELICS broker query to check for EMS federate
The query lets us check if any of the ems-specific interfaces in the storage and
grid federates need to be created. This avoids numerous warnings about
unreceived messages because we no longer sent messages to the ems if it isn't
running.

Using HELICS queries has a major disadvantage: races between federates at
startup. Because we don't know the order federates are started we have to insert
a delay before the query to (hopefully) give enough time for the ems to start if
it is part of the federation. This seems fine for one-off simulations, but as we
start trying to run many simulations within a GA it could become a very annoying
performance hit.

The alternative to this is to pass a flag to let the grid federate know whether
an ems is part of the configuration. This has the disadvantage that it isn't
based on reality, just on what the user says is reality. That said, the result
of an incorrect flag is just a lot of warnings in the best case, but in the
worst case (flag incorrectly says no EMS) the result could be a silent failure
to communicate between the grid federate and the EMS. The result of an error due
to the race condition would be the same silent failure of communication between
the grid and ems federates.

What is implemented here could be easily adapted to use a flag instead of a
HELICS query.